Migrate TLC directory
