+++ apps/dapp/src/components/Pages/Core/DappPages/Borrow/index.tsx
@@ -13,10 +13,10 @@ import templeImg from 'assets/images/newui-images/tokens/temple.png';
 import { formatToken } from 'utils/formatter';
 import { TradeButton } from '../../NewUI/Home';
 import { Popover } from 'components/Popover';
-import Repay from '../../NewUI/TLC/Repay';
-import Borrow from '../../NewUI/TLC/Borrow';
-import Withdraw from '../../NewUI/TLC/Withdraw';
-import Supply from '../../NewUI/TLC/Supply';
+import Repay from './TLC/Repay';
+import Borrow from './TLC/Borrow';
+import Withdraw from './TLC/Withdraw';
+import Supply from './TLC/Supply';
 import { getBigNumberFromString, getTokenInfo } from 'components/Vault/utils';
 import { useNotification } from 'providers/NotificationProvider';
 import { TemplePriceChart } from './Chart';
